+++ src/main/java/com/andersen/test/train/util/CarriageCreator.java
@@ -2,13 +2,14 @@ package com.andersen.test.train.util;
 
 import com.andersen.test.train.model.Carriage;
 import com.andersen.test.train.model.CarriageType;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public class CarriageCreator {
 
-  public Carriage create(CarriageType type, String producer, Double emptyWeight, Integer length,
-      Integer maxPassengerAmount, Double maxValuableWeight, LocalDate productionDate) {
+  public Carriage create(CarriageType type, String producer, BigDecimal emptyWeight, Integer length,
+      Integer maxPassengerAmount, BigDecimal maxValuableWeight, LocalDate productionDate) {
     return Carriage.builder()
         .uid(UUID.randomUUID().toString())
         .type(type)
